Enhance resource list support with fzf
This commit adds functionality to use fzf (if it is installed) as the default output for the resource list, it is a good first pass at #7 that may be expanded on in the future. The defaults for fzf preview (right side/toggle with `?`) make sense for me but could be customised at some point. Making use of `fzf#wrap` means we get a lot of functionality for nothing as it decorates the fzf options with many useful defaults.