Set tiebreak to index for apropos
The length ordering in fzf makes schema.core/defn show up before
clojure.core/defn, which is not what I think is expected.  Using the `index`
tiebreaker ensures ordering is that of what cider sends:

- Current namespace comes first
- Clojure.xxx is second
- Everything else

It may be sensible to also throw in `end`, or perhaps `begin` for ordering too.