Enable docstring preview for ReplantApropos
This commit enables an inline docstring preview using the preview functionality provided by fzf, it does so by writing a temp file made up of the var and docstring as provided by apropos, the fzf preview command is then set to a shell command that greps the temp file and does some munging to produce a readable docstring.

It is likely that this shell command could be improved as I am not great with bash, one point of discussion might be the field range passed to `cut`, is there a more elegant way of cutting and taking all fields from index 2 to end? I want to write something like `cut -d : -f 2-*` but that is not valid, so for now I have used an arbitrary high number. Another point of dicussion might be how portable the shell command is.